refactor(core): extract persist citation helpers under 3000-line ceiling
extraction-persist.ts is now under the file-size ratchet so citation
normalization can live in its own module.

Fixes #3091
